+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [FIX] Fix lost connection in screen streaming
 - [Feature] Change notification icon
 - [FIX] Try to fix impact for connect button
+- [FIX] Atomic write/read manifest storage
 
 # 1.6.7